Update Index.svelte
Summary: Changed text and structure and removed the links at the bottom of the page. I have no confidence that I did anything correctly with styling or commenting out. Sorry. 

Graphic design still has to be done, the site needs to be moved to dominioninstitute.org, an email signup needs to be put on the front page, the "email us" link should probably open an email client, and the image of Dr. K needs to be put in next to his mini-bio.